fix: fixed a wrong denom that was sent to the executeBurn()
