+++ src/components/burner/Burner.tsx
@@ -4,7 +4,6 @@ import type { Asset } from '@/types'
 import type { ChainName } from '@/constants'
 import { useExecuteBurn } from '@/hooks'
 import { BurnerForm } from './BurnerForm'
-import { chains } from 'chain-registry'
 
 interface Props {
   nativeAsset: Asset
@@ -16,8 +15,7 @@ interface Props {
 export const Burner: FC <Props> = ({ chainName, nativeAsset, mintAsset }) => {
   const [burnDisplayAmount, setBurnDisplayAmount] = useState('')
   const amount = Number(burnDisplayAmount) * Math.pow(10, nativeAsset.decimals)
-  const denom = chains.find(({ chain_name: chain }) => chain === chainName)?.fees?.fee_tokens[0].denom ?? ""
-  const executeBurn = useExecuteBurn(chainName, amount, denom)
+  const executeBurn = useExecuteBurn(chainName, amount, nativeAsset.id)
   return (
     <BurnerForm
      disabled={false}
