+++ src/components/complex/ExecuteButton.tsx
@@ -11,10 +11,16 @@ export interface ExecuteButtonProps extends UseExecuteContractResult {
   action: string
   disabled?: boolean
   chainName: ChainName
-  isWalletConnected: boolean
 }
 
-export const ExecuteButton: FC<ExecuteButtonProps> = ({ action, disabled, simulate, sign, broadcast, chainName, }): JSX.Element => {
+export const ExecuteButton: FC<ExecuteButtonProps> = ({
+  action,
+  disabled,
+  simulate,
+  sign,
+  broadcast,
+  chainName
+}): JSX.Element => {
   const [loading, setLoading] = useState(false)
   const modal = useModal()
   const snackbar = useSnackbar()
@@ -22,26 +28,34 @@ export const ExecuteButton: FC<ExecuteButtonProps> = ({ action, disabled, simula
   const execute = (): void => {
     setLoading(true)
     snackbar.open('Simulating', 'info')
-    simulate().then(
-      gas => {
+    simulate()
+      .then((gas) => {
         snackbar.open('Signing', 'info')
-        sign(gas === undefined ? 1_000_000 : gas * 1.2)
-          .then(tx => {
+        sign(Math.ceil(gas === undefined ? 1_000_000 : gas * 1.2))
+          .then((tx) => {
             snackbar.open('Broadcasting...', 'info')
             broadcast(TxRaw.encode(tx).finish())
-              .then(response => modal.open(<TransactionModal deliveryTxResp={response} chainName={chainName} />))
+              .then((response) =>
+                modal.open(
+                <TransactionModal
+                  deliveryTxResp={response}
+                  chainName={chainName} />
+              )
+              )
               .catch(() => snackbar.open('Broadcast Failed', 'error'))
               .finally(() => setLoading(false))
           })
-          .catch(() => {
+          .catch((e) => {
             setLoading(false)
+            console.log('Execute Button Error', e)
             snackbar.open('Request Rejected', 'error')
           })
-      }).catch((e) => {
-      setLoading(false)
-      console.error('Simulation Failure', e)
-      snackbar.open('Simulation Failed', 'error')
-    })
+      })
+      .catch((e) => {
+        setLoading(false)
+        console.error('Simulation Failure', e)
+        snackbar.open('Simulation Failed', 'error')
+      })
   }
 
   return (
@@ -52,7 +66,7 @@ export const ExecuteButton: FC<ExecuteButtonProps> = ({ action, disabled, simula
       size="large"
       sx={{ width: 256 }}
     >
-     {loading ? <CircularProgress color="inherit" size={26} /> : action}
+      {loading ? <CircularProgress color="inherit" size={26} /> : action}
     </Button>
   )
 }
