+++ src/hooks/index.ts
@@ -9,5 +9,4 @@ export * from './useBroadcastTransaction'
 export * from './useExecuteBurn'
 export * from './useFetchChainAssets'
 export * from './useFetchLeaderboard'
-export * from './useFetchFurnaceData'
 export * from './useSimulateTransaction'
